feature: remove certifications
